MM-40114 Rename LegacySidebarHeader to MobileSidebarHeader and remove unused dropdown
