--- components/sidebar/legacy_sidebar_header/index.ts
@@ -1,18 +1,0 @@
-// Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
-// See LICENSE.txt for license information.
-
-import {connect} from 'react-redux';
-
-import {getIsMobileView} from 'selectors/views/browser';
-
-import {GlobalState} from 'types/store';
-
-import LegacySidebarHeader from './legacy_sidebar_header';
-
-function mapStateToProps(state: GlobalState) {
-    return {
-        isMobileView: getIsMobileView(state),
-    };
-}
-
-export default connect(mapStateToProps)(LegacySidebarHeader);
