+++ components/sidebar/mobile_sidebar_header/index.ts
@@ -1,0 +1,5 @@
+// Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
+// See LICENSE.txt for license information.
+
+import MobileSidebarHeader from './mobile_sidebar_header';
+export default MobileSidebarHeader;
