+++ sass/responsive/_mobile.scss
@@ -1502,10 +1502,6 @@
             .sidebar-header-dropdown {
                 display: none;
             }
-
-            .sidebar-header-dropdown__icon {
-                display: none;
-            }
         }
 
         .search__form {
@@ -1596,10 +1592,6 @@
             }
 
             .sidebar-header-dropdown {
-                display: none;
-            }
-
-            .sidebar-header-dropdown__icon {
                 display: none;
             }
         }
